refactor: 优化路径判断，更新
