+++ assets/doc/zh/image_recognition.md
@@ -2,7 +2,9 @@
 
 ## 坐标系统
 
-项目中的坐标原点为 **游戏窗口客户区（client area）的左上角**。
+图片的坐标原点为 **游戏窗口客户区的左上角**。
+
+find_element 返回的坐标是当前游戏窗口客户区坐标，而硬编码的坐标基于 2560×1440 分辨率。因此，在使用find_element返回的坐标时需要**缩放**。
 
 ## 图片查找路径机制
 
@@ -10,6 +12,14 @@
 |----------|----------|
 | `zh_cn` (中文) | `["zh_cn", "en", "share"]` |
 | `en` (英文) | `["en", "share"]` |
+
+查找到首个匹配文件即停止搜索。
+
+## 图像路径更新机制
+
+启动时根据游戏语言初始化图片搜索路径。
+
+当 `dark` 模板匹配失败会尝试 `default` 模板，若成功则移除 `dark` 路径。
 
 ## 图像识别
 
@@ -31,11 +41,9 @@ e.g.
 
 ### 图片制作准则
 
-1. **基准分辨率**：所有图片资源steam端全屏 **2560×1440 (1440p)** 分辨率下截取
+1. **基准分辨率**：所有图片资源steam端**全屏 2560×1440** 分辨率下截取
 2. **文件命名**：含黑幕的完整截图使用 `assets` 后缀，只有核心图像的截图不使用 `assets` 后缀
 3. **优先使用含黑幕的完整截图**：便于系统自动计算有效区域，适应UI变化
-
-分享一个制作链接：[figma](https://www.figma.com/design/DEGOtWVZ5rLzXzX5jXd54h/image?node-id=0-1&t=f9i0jD8xf5fCvM2x-1)
 
 ### 参数
