+++ utils/__init__.py
@@ -1,3 +1,1 @@
-from .path_manager import path_manager
 
-pic_path = ["default/en", "default/share"]
